combined_areas {noncensus}	R Documentation

Combined Statistical Areas (CSAs)

Description

The U.S. Census Bureau groups counties into CSAs primarily based on county population. NOTE: Not all counties are members of a CSA. For a detailed description, Wikipedia has an excellent discussion: http://en.wikipedia.org/wiki/Combined_Statistical_Area. Also, the following map from Wikipedia is excellent to visualize the areas: http://upload.wikimedia.org/wikipedia/commons/7/7b/Combined_statistical_areas_of_the_United_States_and_Puerto_Rico.gif

Usage

data(combined_areas)

Format

A data frame with 166 rows and 2 variables.

[Package noncensus version 0.1 Index]