corebased_areas {noncensus}	R Documentation

Core-based Statistical Area (CBSAs)

Description

The U.S. Census Bureau groups counties into CBSAs primarily based on county population. NOTE: Not all counties are members of a CBSA. For a detailed description, Wikipedia has an excellent discussion: http://en.wikipedia.org/wiki/Core_Based_Statistical_Area. Also, the following map from Wikipedia is excellent to visualize the areas: http://upload.wikimedia.org/wikipedia/commons/7/7b/Combined_statistical_areas_of_the_United_States_and_Puerto_Rico.gif

Usage

data(corebased_areas)

Format

A data frame with 917 rows and 4 variables.

[Package noncensus version 0.1 Index]