counties {noncensus}	R Documentation

Data for U.S. Counties and County-Equivalent Entities

Description

Data containing state and county FIPS codes for U.S. counties and county-equivalent entities (CEE) along with county-level demographic data. The CEE includes non-state locations, such as Puerto Rico (PR) and Guam (GU).

Usage

data(counties)

Format

A data frame with 3235 rows and 6 variables

Details

county_name. County Name and Legal/Statistical Area Description

state. State Postal Code

state_fips. State FIPS Code

county_fips. County FIPS Code

fips_class. FIPS Class Code

CSA. Combined Statistical Area

CBSA. Core-based Statistical Area

population. County population from 2010 Census

The U.S. Census Bureau groups counties into CSAs and CBSAs primarily based on county population. We provide listings of both in combined_areas and corebased_areas.

For a detailed description, Wikipedia has excellent discussions of both areas: http://en.wikipedia.org/wiki/Combined_Statistical_Area and http://en.wikipedia.org/wiki/Core_Based_Statistical_Area. Also, the following map from Wikipedia is excellent to visualize the areas: http://upload.wikimedia.org/wikipedia/commons/7/7b/Combined_statistical_areas_of_the_United_States_and_Puerto_Rico.gif

NOTE: Not all counties are members of a CSA or CBSA.

The following details about FIPS Class Codes have been blatantly taken from the Census Bureau's website:

H1. Identifies an active county or statistically equivalent entity that does not qualify under subclass C7 or H6.

H4. Identifies a legally defined inactive or nonfunctioning county or statistically equivalent entity that does not qualify under subclass H6.

H5. Identifies census areas in Alaska, a statistical county equivalent entity.

H6. Identifies a county or statistically equivalent entity that is areally coextensive or governmentally consolidated with an incorporated place, part of an incorporated place, or a consolidated city.

C7: Identifies an incorporated place that is an independent city; that is, it also serves as a county equivalent because it is not part of any county, and a minor civil division (MCD) equivalent because it is not part of any MCD.

For more details, see: http://www.census.gov/geo/reference/codes/cou.html

[Package noncensus version 0.1 Index]