noncensus {noncensus}	R Documentation

U.S. Census Region and Demographic Data

Description

The R package noncensus provides a collection of various regional information determined by the U.S. Census Bureau along with demographic data. We have also included scripts to download, process, and load the original data from their sources.

[Package noncensus version 0.1 Index]