states {noncensus}	R Documentation

Demographic Data and Census Regions of U.S. States and Territories

Description

A dataset containing demographic information and the census regions of each U.S. state as defined by the U.S. Census Bureau. Also included are the U.S. territories, such as Puerto Rico and Guam.

Usage

data(states)

Format

A data frame with 56 rows and 7 variables

Details

The variables included are:

state. State abbreviation

name. State name

region. Region as defined by the U.S. Census Bureau

division. Subregion as defined by the U.S. Census Bureau

capital. Capital city.

area. Land area in square miles

population. Population from 2010 Census

The U.S. is divided into four regions:

Midwest

Northeast

South

West

Within each region, states are further partitioned into divisions. For more details about census regions, see: http://en.wikipedia.org/wiki/List_of_regions_of_the_United_States#Census_Bureau-designated_regions_and_divisions

Much of the state data was extracted from http://www.census.gov/popest/data/state/totals/2013/index.html

[Package noncensus version 0.1 Index]