zip_codes {noncensus}	R Documentation

Data for U.S. Cities by Zip Code

Description

This data set considers each zip code throughout the U.S. and provides additional information, including the city and state, latitude and longitude, and the FIPS code for the corresponding county.

Usage

data(zip_codes)

Format

A data frame with 43524 rows and 6 variables

Details

The ZIP code data was obtained from version 1.0 of the zipcode package on CRAN. The county FIPS codes were obtained by querying the FIPS code from each zip's latitude and longitude via the FCC's Census Block Conversions API. For details regarding the API, see http://www.fcc.gov/developers/census-block-conversions-api.

zip. U.S. ZIP (postal) code

city. ZIP code's city

state. ZIP code's state

latitude. ZIP code's latitude

longitude. ZIP code's longitude

fips. County FIPS Code

The FIPS codes are useful for mapping ZIP codes and cities to counties in the counties data set.

Fun fact: ZIP is an acronym for "Zone Improvement Plan."

[Package noncensus version 0.1 Index]